requirements.txt
evdev
HAP-python[QRCode]
nicegui
pillow
pyyaml
qrcode
scrypt
watchdog
